fix: Resolve compilation errors
- Fixed @Composable context issue in LoadingComponents.kt
- Fixed type mismatch in ImageCache.kt (Int vs Long)
- Application now compiles successfully
- All icons are properly imported and working

+++ app/src/main/java/com/abdapps/scandocs/utils/ImageCache.kt
@@ -226,8 +226,8 @@ class ImageCache(private val context: Context) {
         return CacheStats(
             memorySize = memoryCache.size(),
             memoryMaxSize = memoryCache.maxSize(),
-            memoryHitCount = memoryCache.hitCount(),
-            memoryMissCount = memoryCache.missCount()
+            memoryHitCount = memoryCache.hitCount().toLong(),
+            memoryMissCount = memoryCache.missCount().toLong()
         )
     }
 }
